Clear latched MCP23xxx interrupts during setup, fix uint8_t truncation
- MCP23008/MCP23S08: Read GPIO register before attaching ISR to clear
  any latched interrupt from a previous MCU run (chip may stay powered)
- MCP23017/MCP23S17: Read both GPIOA and GPIOB to clear latched state
- PI4IOE5V6408: Add explicit static_cast<uint8_t> for ~mode_mask_ to
  avoid implicit int-to-uint8_t truncation

+++ esphome/components/pi4ioe5v6408/pi4ioe5v6408.cpp
@@ -162,7 +162,7 @@ bool PI4IOE5V6408Component::write_gpio_modes_() {
   // Enable interrupts for input pins when interrupt pin is configured
   // (input pins have mode_mask_ bit cleared)
   if (this->interrupt_pin_ != nullptr &&
-      !this->write_byte(PI4IOE5V6408_REGISTER_INTERRUPT_ENABLE_MASK, ~this->mode_mask_)) {
+      !this->write_byte(PI4IOE5V6408_REGISTER_INTERRUPT_ENABLE_MASK, static_cast<uint8_t>(~this->mode_mask_))) {
     this->status_set_warning(LOG_STR("Failed to write interrupt enable mask"));
     return false;
   }
